+++ src/api/projects/edit.rs
@@ -88,6 +88,7 @@ pub struct EditProject<'a> {
     resolve_outdated_diff_discussions: Option<bool>,
     /// Whether the container registry is enabled or not.
     #[builder(default)]
+    #[deprecated(note = "use `container_registry_access_level` instead")]
     container_registry_enabled: Option<bool>,
     /// The expiration policy for containers.
     #[builder(default)]
@@ -262,6 +263,7 @@ impl<'a> EditProject<'a> {
 
 impl<'a> EditProjectBuilder<'a> {
     /// Add a tag.
+    #[deprecated(note = "use `topic` instead")]
     pub fn tag<T>(&mut self, tag: T) -> &mut Self
     where
         T: Into<Cow<'a, str>>,
@@ -273,6 +275,7 @@ impl<'a> EditProjectBuilder<'a> {
     }
 
     /// Add multiple tags.
+    #[deprecated(note = "use `topics` instead")]
     pub fn tags<I, T>(&mut self, iter: I) -> &mut Self
     where
         I: Iterator<Item = T>,
@@ -360,10 +363,6 @@ impl<'a> Endpoint for EditProject<'a> {
             .push_opt(
                 "resolve_outdated_diff_discussions",
                 self.resolve_outdated_diff_discussions,
-            )
-            .push_opt(
-                "container_registry_enabled",
-                self.container_registry_enabled,
             )
             .push_opt("shared_runners_enabled", self.shared_runners_enabled)
             .push_opt("visibility", self.visibility)
@@ -465,7 +464,11 @@ impl<'a> Endpoint for EditProject<'a> {
                 .push_opt("merge_requests_enabled", self.merge_requests_enabled)
                 .push_opt("jobs_enabled", self.jobs_enabled)
                 .push_opt("wiki_enabled", self.wiki_enabled)
-                .push_opt("snippets_enabled", self.snippets_enabled);
+                .push_opt("snippets_enabled", self.snippets_enabled)
+                .push_opt(
+                    "container_registry_enabled",
+                    self.container_registry_enabled,
+                );
         }
 
         params.into_body()
@@ -896,6 +899,7 @@ mod tests {
     }
 
     #[test]
+    #[allow(deprecated)]
     fn endpoint_container_registry_enabled() {
         let endpoint = ExpectedUrl::builder()
             .method(Method::PUT)
@@ -1468,6 +1472,7 @@ mod tests {
     }
 
     #[test]
+    #[allow(deprecated)]
     fn endpoint_tag_list() {
         let endpoint = ExpectedUrl::builder()
             .method(Method::PUT)